+++ src/neo/core/config/routes.config.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * Neo Framework
+ *
+ * @link https://neo-framework.github.io
+ * @copyright Copyright (c) 2016-2017 YouniS Bensalah <younis.bensalah@gmail.com>
+ * @license MIT
+ */
+
+return [
+    'routes' => [
+
+        //
+
+    ]
+];